pyproject.toml
[tool.poetry]
name = "diamond-atari"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
include = [
    "config/**",
    "dweam.toml",
    "thumbnails/**",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
gymnasium = {version = "0.29.1", extras = ["accept-rom-license", "atari"]}
h5py = "3.11.0"
huggingface-hub = "0.17.2"
hydra-core = "1.3"
numpy = "1.26.2"
opencv-python = "4.8.0.74"
pillow = "10.3.0"
pygame = "2.5.2"
torch = {version = "2.4.1", source = "pytorch-cuda-121"}
torchvision = {version = "0.19.1", source = "pytorch-cuda-121"}
torcheval = "0.0.7"
tqdm = "4.66.4"
wandb = "0.17.0"
ale-py = "0.8.1"

[tool.poetry.group.dev.dependencies]
tomli-w = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cuda-121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"
